// src/components/rating/enum.ts
/**
 * CSS Classes
 */
export enum CssClass {
    Rating = 'osui-rating',
    IsEdit = 'is-edit',
    Filled = 'rating-item-filled',
    Empty = 'rating-item-empty',
    RatingInput = 'rating-input',
    RatingItem = 'rating-item',
    WCAGHideText = 'wcag-hide-text',
}

/**
 * Callbacks eventName
 */
export enum Events {
    OnSelected = 'OnSelected',
}

/**
 * Properties
 */
export enum Properties {
    IsEdit = 'IsEdit',
    MaxRatingScale = 100,
    MinRatingScale = 0,
    RatingScale = 'RatingScale',
    RatingValue = 'RatingValue',
    Size = 'Size',
}